requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
mysqlclient>=2.2.0
python-decouple>=3.8
requests>=2.31.0
Pillow>=10.0.0
